requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
flask==2.2.2
frozenlist==1.3.1
future==0.18.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0 ; python_version < '3.10'
itsdangerous==2.1.2
jinja2==3.1.2
line-bot-sdk==2.3.0
markupsafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
